Make test util functions return trait object
