+++ src/lib.rs
@@ -4,3 +4,5 @@ pub mod elements;
 pub mod graph_agdb;
 pub(crate) mod graph_traits;
 pub mod nodetype;
+
+mod utils;
